requirements.txt
numpy
pandas
spacy
scikit-learn
matplotlib
scipy
requests
pillow
tqdm
tensorflow
PyYAML
faiss-cpu
streamlit
Pillow
-e .